+++ python/x402/http/facilitator_client.py
@@ -7,9 +7,7 @@ implementations for communicating with remote facilitator services.
 from __future__ import annotations
 
 import json
-from collections.abc import Callable
-from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Any, Protocol
+from typing import TYPE_CHECKING, Any
 
 from ..schemas import (
     PaymentPayload,
@@ -19,188 +17,30 @@ from ..schemas import (
     VerifyResponse,
 )
 from ..schemas.v1 import PaymentPayloadV1, PaymentRequirementsV1
-from .constants import DEFAULT_FACILITATOR_URL
+from .facilitator_client_base import (
+    AuthHeaders,
+    AuthProvider,
+    CreateHeadersAuthProvider,
+    FacilitatorClient,
+    FacilitatorClientSync,
+    FacilitatorConfig,
+    HTTPFacilitatorClientBase,
+)
 
 if TYPE_CHECKING:
     import httpx
 
-
-# ============================================================================
-# Auth Provider Protocol
-# ============================================================================
-
-
-@dataclass
-class AuthHeaders:
-    """Authentication headers for facilitator endpoints."""
-
-    verify: dict[str, str] = field(default_factory=dict)
-    settle: dict[str, str] = field(default_factory=dict)
-    supported: dict[str, str] = field(default_factory=dict)
-
-
-class AuthProvider(Protocol):
-    """Generates authentication headers for facilitator requests."""
-
-    def get_auth_headers(self) -> AuthHeaders:
-        """Get authentication headers for each endpoint."""
-        ...
-
-
-class CreateHeadersAuthProvider:
-    """AuthProvider that wraps a create_headers callable.
-
-    Adapts the dict-style create_headers function (as used by CDP SDK)
-    to the AuthProvider protocol.
-    """
-
-    def __init__(self, create_headers: Callable[[], dict[str, dict[str, str]]]) -> None:
-        self._create_headers = create_headers
-
-    def get_auth_headers(self) -> AuthHeaders:
-        """Get authentication headers by calling the create_headers function."""
-        result = self._create_headers()
-        return AuthHeaders(
-            verify=result.get("verify", {}),
-            settle=result.get("settle", {}),
-            supported=result.get("supported", result.get("list", {})),
-        )
-
-
-# ============================================================================
-# FacilitatorClient Protocols
-# ============================================================================
-
-
-class FacilitatorClient(Protocol):
-    """Protocol for async facilitator clients.
-
-    Used by x402ResourceServer (async) to verify/settle payments.
-    """
-
-    async def verify(
-        self,
-        payload: PaymentPayload | PaymentPayloadV1,
-        requirements: PaymentRequirements | PaymentRequirementsV1,
-    ) -> VerifyResponse:
-        """Verify a payment. Returns VerifyResponse with is_valid=False on failure."""
-        ...
-
-    async def settle(
-        self,
-        payload: PaymentPayload | PaymentPayloadV1,
-        requirements: PaymentRequirements | PaymentRequirementsV1,
-    ) -> SettleResponse:
-        """Settle a payment. Returns SettleResponse with success=False on failure."""
-        ...
-
-    def get_supported(self) -> SupportedResponse:
-        """Get supported payment kinds (sync - used during initialization)."""
-        ...
-
-
-class FacilitatorClientSync(Protocol):
-    """Protocol for sync facilitator clients.
-
-    Used by x402ResourceServerSync (sync) to verify/settle payments.
-    """
-
-    def verify(
-        self,
-        payload: PaymentPayload | PaymentPayloadV1,
-        requirements: PaymentRequirements | PaymentRequirementsV1,
-    ) -> VerifyResponse:
-        """Verify a payment. Returns VerifyResponse with is_valid=False on failure."""
-        ...
-
-    def settle(
-        self,
-        payload: PaymentPayload | PaymentPayloadV1,
-        requirements: PaymentRequirements | PaymentRequirementsV1,
-    ) -> SettleResponse:
-        """Settle a payment. Returns SettleResponse with success=False on failure."""
-        ...
-
-    def get_supported(self) -> SupportedResponse:
-        """Get supported payment kinds."""
-        ...
-
-
-# ============================================================================
-# Configuration
-# ============================================================================
-
-
-@dataclass
-class FacilitatorConfig:
-    """Configuration for HTTP facilitator client."""
-
-    url: str = DEFAULT_FACILITATOR_URL
-    timeout: float = 30.0
-    http_client: Any = None  # Optional httpx.Client or httpx.AsyncClient
-    auth_provider: AuthProvider | None = None
-    identifier: str | None = None
-
-
-# ============================================================================
-# Base HTTP Facilitator Client (Shared Logic)
-# ============================================================================
-
-
-class _HTTPFacilitatorClientBase:
-    """Base class with shared logic for HTTP facilitator clients."""
-
-    def __init__(self, config: FacilitatorConfig | dict[str, Any] | None = None) -> None:
-        """Create HTTP facilitator client.
-
-        Args:
-            config: Optional configuration. Accepts either:
-                - FacilitatorConfig dataclass (recommended)
-                - Dict with 'url' and optional 'create_headers'
-                - None (uses defaults)
-        """
-        # Handle dict-style config
-        if isinstance(config, dict):
-            url = config.get("url", DEFAULT_FACILITATOR_URL)
-            create_headers = config.get("create_headers")
-            auth_provider = CreateHeadersAuthProvider(create_headers) if create_headers else None
-
-            self._url = url.rstrip("/")
-            self._timeout = 30.0
-            self._auth_provider = auth_provider
-            self._identifier = self._url
-            self._http_client = None
-            self._owns_client = True
-        else:
-            # Handle FacilitatorConfig dataclass or None
-            config = config or FacilitatorConfig()
-
-            self._url = config.url.rstrip("/")
-            self._timeout = config.timeout
-            self._auth_provider = config.auth_provider
-            self._identifier = config.identifier or self._url
-            self._http_client = config.http_client
-            self._owns_client = config.http_client is None
-
-    @property
-    def url(self) -> str:
-        """Get facilitator URL."""
-        return self._url
-
-    @property
-    def identifier(self) -> str:
-        """Get facilitator identifier."""
-        return self._identifier
-
-    @staticmethod
-    def _to_json_safe(obj: Any) -> Any:
-        """Convert object to JSON-safe format (handles bigints)."""
-        return json.loads(
-            json.dumps(
-                obj,
-                default=lambda x: str(x) if isinstance(x, int) and x > 2**53 else x,
-            )
-        )
+# Re-export for external use
+__all__ = [
+    "HTTPFacilitatorClient",
+    "HTTPFacilitatorClientSync",
+    "FacilitatorConfig",
+    "FacilitatorClient",
+    "FacilitatorClientSync",
+    "AuthProvider",
+    "AuthHeaders",
+    "CreateHeadersAuthProvider",
+]
 
 
 # ============================================================================
@@ -208,7 +48,7 @@ class _HTTPFacilitatorClientBase:
 # ============================================================================
 
 
-class HTTPFacilitatorClient(_HTTPFacilitatorClientBase):
+class HTTPFacilitatorClient(HTTPFacilitatorClientBase):
     """Async HTTP-based facilitator client.
 
     Communicates with remote x402 facilitator services over HTTP using
@@ -317,21 +157,17 @@ class HTTPFacilitatorClient(_HTTPFacilitatorClientBase):
         """
         # Use sync client for initialization (called from sync initialize())
         with self._get_sync_client() as client:
-            headers = {"Content-Type": "application/json"}
-
-            if self._auth_provider:
-                auth = self._auth_provider.get_auth_headers()
-                headers.update(auth.supported)
-
-            response = client.get(f"{self._url}/supported", headers=headers)
+            response = client.get(
+                f"{self._url}/supported",
+                headers=self._get_supported_headers(),
+            )
 
             if response.status_code != 200:
                 raise ValueError(
                     f"Facilitator get_supported failed ({response.status_code}): {response.text}"
                 )
 
-            data = response.json()
-            return SupportedResponse.model_validate(data)
+            return SupportedResponse.model_validate(response.json())
 
     # =========================================================================
     # Bytes-Based Methods (Network Boundary)
@@ -397,30 +233,18 @@ class HTTPFacilitatorClient(_HTTPFacilitatorClientBase):
     ) -> VerifyResponse:
         """Internal verify via HTTP (async)."""
         client = self._get_async_client()
-
-        headers = {"Content-Type": "application/json"}
-
-        if self._auth_provider:
-            auth = self._auth_provider.get_auth_headers()
-            headers.update(auth.verify)
-
-        request_body = {
-            "x402Version": version,
-            "paymentPayload": self._to_json_safe(payload_dict),
-            "paymentRequirements": self._to_json_safe(requirements_dict),
-        }
+        request_body = self._build_request_body(version, payload_dict, requirements_dict)
 
         response = await client.post(
             f"{self._url}/verify",
-            headers=headers,
+            headers=self._get_verify_headers(),
             json=request_body,
         )
 
         if response.status_code != 200:
             raise ValueError(f"Facilitator verify failed ({response.status_code}): {response.text}")
 
-        data = response.json()
-        return VerifyResponse.model_validate(data)
+        return VerifyResponse.model_validate(response.json())
 
     async def _settle_http(
         self,
@@ -430,30 +254,18 @@ class HTTPFacilitatorClient(_HTTPFacilitatorClientBase):
     ) -> SettleResponse:
         """Internal settle via HTTP (async)."""
         client = self._get_async_client()
-
-        headers = {"Content-Type": "application/json"}
-
-        if self._auth_provider:
-            auth = self._auth_provider.get_auth_headers()
-            headers.update(auth.settle)
-
-        request_body = {
-            "x402Version": version,
-            "paymentPayload": self._to_json_safe(payload_dict),
-            "paymentRequirements": self._to_json_safe(requirements_dict),
-        }
+        request_body = self._build_request_body(version, payload_dict, requirements_dict)
 
         response = await client.post(
             f"{self._url}/settle",
-            headers=headers,
+            headers=self._get_settle_headers(),
             json=request_body,
         )
 
         if response.status_code != 200:
             raise ValueError(f"Facilitator settle failed ({response.status_code}): {response.text}")
 
-        data = response.json()
-        return SettleResponse.model_validate(data)
+        return SettleResponse.model_validate(response.json())
 
 
 # ============================================================================
@@ -461,7 +273,7 @@ class HTTPFacilitatorClient(_HTTPFacilitatorClientBase):
 # ============================================================================
 
 
-class HTTPFacilitatorClientSync(_HTTPFacilitatorClientBase):
+class HTTPFacilitatorClientSync(HTTPFacilitatorClientBase):
     """Sync HTTP-based facilitator client.
 
     Communicates with remote x402 facilitator services over HTTP using
@@ -561,21 +373,17 @@ class HTTPFacilitatorClientSync(_HTTPFacilitatorClientBase):
         """
         client = self._get_client()
 
-        headers = {"Content-Type": "application/json"}
-
-        if self._auth_provider:
-            auth = self._auth_provider.get_auth_headers()
-            headers.update(auth.supported)
-
-        response = client.get(f"{self._url}/supported", headers=headers)
+        response = client.get(
+            f"{self._url}/supported",
+            headers=self._get_supported_headers(),
+        )
 
         if response.status_code != 200:
             raise ValueError(
                 f"Facilitator get_supported failed ({response.status_code}): {response.text}"
             )
 
-        data = response.json()
-        return SupportedResponse.model_validate(data)
+        return SupportedResponse.model_validate(response.json())
 
     # =========================================================================
     # Bytes-Based Methods (Network Boundary)
@@ -641,30 +449,18 @@ class HTTPFacilitatorClientSync(_HTTPFacilitatorClientBase):
     ) -> VerifyResponse:
         """Internal verify via HTTP."""
         client = self._get_client()
-
-        headers = {"Content-Type": "application/json"}
-
-        if self._auth_provider:
-            auth = self._auth_provider.get_auth_headers()
-            headers.update(auth.verify)
-
-        request_body = {
-            "x402Version": version,
-            "paymentPayload": self._to_json_safe(payload_dict),
-            "paymentRequirements": self._to_json_safe(requirements_dict),
-        }
+        request_body = self._build_request_body(version, payload_dict, requirements_dict)
 
         response = client.post(
             f"{self._url}/verify",
-            headers=headers,
+            headers=self._get_verify_headers(),
             json=request_body,
         )
 
         if response.status_code != 200:
             raise ValueError(f"Facilitator verify failed ({response.status_code}): {response.text}")
 
-        data = response.json()
-        return VerifyResponse.model_validate(data)
+        return VerifyResponse.model_validate(response.json())
 
     def _settle_http(
         self,
@@ -674,27 +470,15 @@ class HTTPFacilitatorClientSync(_HTTPFacilitatorClientBase):
     ) -> SettleResponse:
         """Internal settle via HTTP."""
         client = self._get_client()
-
-        headers = {"Content-Type": "application/json"}
-
-        if self._auth_provider:
-            auth = self._auth_provider.get_auth_headers()
-            headers.update(auth.settle)
-
-        request_body = {
-            "x402Version": version,
-            "paymentPayload": self._to_json_safe(payload_dict),
-            "paymentRequirements": self._to_json_safe(requirements_dict),
-        }
+        request_body = self._build_request_body(version, payload_dict, requirements_dict)
 
         response = client.post(
             f"{self._url}/settle",
-            headers=headers,
+            headers=self._get_settle_headers(),
             json=request_body,
         )
 
         if response.status_code != 200:
             raise ValueError(f"Facilitator settle failed ({response.status_code}): {response.text}")
 
-        data = response.json()
-        return SettleResponse.model_validate(data)
+        return SettleResponse.model_validate(response.json())
